Legacy Plan Support (#281)
Added a compatibility layer users can enable if they're changing from v3 to v4.

In v3 and below, GUIDs were generated for plans if they weren't defined by the user. In v4, we disabled this "feature" because it makes debugging, documentation, and upgrades overly complex. Instead, if you enable the compatibility layer the broker will intercept all calls to the backing-services and check if they're one of the legacy plans we used to define. If they are, then we reject the request with a message to upgrade the service like the following: `cf update-service pcf_sb_1_1536963604087466937 -p nearline`

Tested by @mattysweeps and I in a PCF installation he owned, started with tile 3.x, created all services, upgraded to 4.0 service broker, validated error messages were correct, then upgraded the plans, disabled the compatibility layer and validated the new broker would provision plans.