Cleanup of CloudSQL logic and fixup LastOperation
This commit cleans up the way CloudSQL was halding operations.
Rather than keeping a separate table for operations, they now get stored
with a state tag and operation name as part of the instance information.
This allows us to comply with the OSB spec better and approach things
more transactionally with fewer failure states.

Storing the state information in the instance with a well-defined
state machine also means we can reason about the state of the service
in the gcp_service_broker implementation without resorting to
separate function calls that had to be implemented on BrokerHelpers.

Finally, it has allowed us to do away with the last special case that
our DAO had to deal with and the entire DAO entry for CloudOperations
which was about a 600 line savings.