CloudSQL schemas (#310)
I've gone and defined the schemas before changing the code for `broker.go` so you can compare the two.

I added the ability to do assertions in the broker variables. The set variables won't affect anything when run, but we do need them in order to maintain the same functionality we had before.

If the CloudSQL API behaves correctly, then we shouldn't need the assertion at all.

Based on all the transformation I've been doing, I wrote down guidelines for future services we (and others) build out.

There is a "breaking" change in this code when combined with the JSONSchema code. The hard drive constants we were provisioning with before weren't in any way valid, I suspect the CloudSQL API was seeing an invalid value and defaulting back to SSD, but if that's the case then users that entered "hdd" thinking it was the other option would instead be getting the more expensive SSD as well. Therefore I'm considering it a bug fix instead.

I also have some hesitations that we're not doing PostgreSQL HA correctly, but that's another story.