mark built-in services as not-free (#408)
Fixes #49. This PR marks all built-in services as non-free. It doesn't attempt to fetch pricing information, but we can at least indicate that services will incur cost.

An additional test is added for plan updating to be false for built-ins because we have no intention of opening that can of worms and risking data loss.