ci: Adopt Cloud Build and remove Concourse
This change removes Concourse pipelines in favor of Cloud Build
configurations. It introduces a configuration for unit+integration
testing, as well as a configuration for that testing plus releasing to a
GCS bucket.

Cloud Secrets manager is required to store the service accont key used
by integration tests and is covered in `ci/README.md`.